requirements.txt
Flask==2.3.3
python-docx==0.8.11
Werkzeug==2.3.7
lxml==4.9.3
anthropic==0.8.1
openai==1.6.1
gunicorn==21.2.0